Formatted the study title heading in StudyMonitor.vue
New function that generates a string with study titles of different languages. If a `studytitle` key has an empty value, it's not included in the returned string. The result is that, if a study doesn't define the study titles of other languages, the displayed title won't be e.g. "en: Study, sv: , it: " but rather "en: Study".